Ifdefs to protect tests not supported by enabled scalar types.
Modified Anasazi/Tpetra MVOPTester test to be smaller and use instantiation-aware ifdefs. Removed some warnings in the Anasazi/Tpetra adaptor.

=============================
Build/Test Cases Summary
=============================

Wed Mar  3 17:23:46 EST 2010

Enabled Packages: TrilinosFramework, Anasazi, Belos, Kokkos, Tpetra
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=451,notpassed=0 (27.4119651357 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=475,notpassed=0 (7.12029486497 min)

Local commits for this build/test group:
----------------------------------------
432ed44 Ifdefs to protect tests not supported by enabled scalar types.
9a6144e Half-baked explicit instantiation previously done for Tpetra Matrix IO sub-package was causing problems in the context of new explicit instantiation scheme. Put Matrix IO on proper explicit instantiation.
5128f97 Added explicit instantiation for CrsMatrix, CrsMatrixMultiplyOp and CrsMatrixSolveOp classes.
99654dc Modified FindTPLCUDA.cmake to properly add libraries to build system, so they show up in Makefiles. Also, explicitly/globally add cublas libs to Kokkos TPLs.

+++ packages/anasazi/tpetra/src/AnasaziTpetraAdapter.hpp
@@ -80,7 +80,7 @@ namespace Anasazi {
 #ifdef HAVE_TPETRA_DEBUG
       TEST_FOR_EXCEPTION( *std::min_element(index.begin(),index.end()) < 0, std::runtime_error,
           "Anasazi::MultiVecTraits<Scalar,Tpetra::MultiVector>::Clone(mv,index): indices must be >= zero.");
-      TEST_FOR_EXCEPTION( *std::max_element(index.begin(),index.end()) >= mv.getNumVectors(), std::runtime_error,
+      TEST_FOR_EXCEPTION( (size_t)*std::max_element(index.begin(),index.end()) >= mv.getNumVectors(), std::runtime_error,
           "Anasazi::MultiVecTraits<Scalar,Tpetra::MultiVector>::Clone(mv,index): indices must be < mv.getNumVectors().");
 #endif
       for (typename std::vector<int>::size_type j=1; j<index.size(); ++j) {
@@ -101,7 +101,7 @@ namespace Anasazi {
 #ifdef HAVE_TPETRA_DEBUG
       TEST_FOR_EXCEPTION( *std::min_element(index.begin(),index.end()) < 0, std::invalid_argument,
           "Anasazi::MultiVecTraits<Scalar,Tpetra::MultiVector>::CloneView(mv,index): indices must be >= zero.");
-      TEST_FOR_EXCEPTION( *std::max_element(index.begin(),index.end()) >= mv.getNumVectors(), std::invalid_argument,
+      TEST_FOR_EXCEPTION( (size_t)*std::max_element(index.begin(),index.end()) >= mv.getNumVectors(), std::invalid_argument,
           "Anasazi::MultiVecTraits<Scalar,Tpetra::MultiVector>::CloneView(mv,index): indices must be < mv.getNumVectors().");
 #endif
       for (typename std::vector<int>::size_type j=1; j<index.size(); ++j) {
@@ -122,7 +122,7 @@ namespace Anasazi {
 #ifdef HAVE_TPETRA_DEBUG
       TEST_FOR_EXCEPTION( *std::min_element(index.begin(),index.end()) < 0, std::invalid_argument,
           "Anasazi::MultiVecTraits<Scalar,Tpetra::MultiVector>::CloneView(mv,index): indices must be >= zero.");
-      TEST_FOR_EXCEPTION( *std::max_element(index.begin(),index.end()) >= mv.getNumVectors(), std::invalid_argument,
+      TEST_FOR_EXCEPTION( (size_t)*std::max_element(index.begin(),index.end()) >= mv.getNumVectors(), std::invalid_argument,
           "Anasazi::MultiVecTraits<Scalar,Tpetra::MultiVector>::CloneView(mv,index): indices must be < mv.getNumVectors().");
 #endif
       for (typename std::vector<int>::size_type j=1; j<index.size(); ++j) {
